Create Number_Guessing_Game.java
Number Guessing Game
This is a simple Java-based Number Guessing Game where the player has to guess a randomly generated number within a specified range. The game provides the player with a limited number of attempts to guess the correct number. After each incorrect guess, the game gives a hint whether the correct number is higher or lower than the player's guess.

Key Features:

The number to be guessed is randomly generated within a range of 1 to 100.
The player has a maximum of 10 attempts to guess the correct number.
The game informs the player if the guessed number is higher or lower than the target number.
After running out of attempts or guessing correctly, the player is given the option to play again.
How to Play:

Run the game.
The game will prompt you to guess a number between 1 and 100.
Enter your guess and the game will provide feedback.
Continue guessing until you either find the correct number or run out of attempts.
After the game ends, you can choose to play again or exit.